using Naive Approach

+++ DATA STRUCTURES/Array/Mean of Range in An Array/code_1.cpp
@@ -23,7 +23,7 @@ void getRangeMean(vector<int>arr, vector<Query> List) {
             sum += arr[i];
             count++;
         }
-        double mean = sum / count;
+        double mean = double(sum) / double(count);
         cout << "\nMean for ( " << L << " , " << R << " )\t:\t" << mean << "\n";
     }
     cout << endl;
